Support xCode 13 w/Cocoapods 1.11
This is probably not necessary to merge, but I noticed when I pulled and
built it made a bunch of changes because my Cocoapods version was newer
than the one we have checked in.

I not smert enough to know what's up with this, so I'm putting it as a
PR into your PR so you can ignore it (or merge it if it makes sense? I
dunno!)